chore(protocols): define NeighborStore 4-method protocol
Narrow neighbor_candidates.py type from SessionQueryRuntimeStore (~20 methods)
to NeighborStore (4 methods actually used):
  - resolve_id
  - get
  - list_summaries_by_query
  - search_summary_hits

Updates:
- Add NeighborStore protocol to protocols.py with Consumed-by comment
- Import NeighborStore in neighbor_candidates.py, remove local TYPE_CHECKING
  definition
- Update _ArchiveNeighborRuntime in api/archive.py to return SessionId from
  resolve_id (was str, now properly typed) with no cast needed
- Regenerate topology-target.yaml after protocol addition

_ArchiveNeighborRuntime is already <100 lines; no further trimming needed.
Behavior unchanged (existing tests); mypy --strict green.

Ref polylogue-a7xr.12
